New Kepler solver is now working very nicely for parabolas but I went back to using the older method for ellipses and hyperbolas. It turns out that the universal method is not very robust for hyperbolas and also tends to require a few more iterations than I'd like. Consider the new Kepler solver to be "working" at this time. I am not always getting the same answers as before implementing the parabola solver but I suspect that's because I was wrong before and right now.
Unfortunately the derivatives, while looking very nice in the fort, still don't work. Oh well.

Also I made a change to NSGA-II: when a solution is placed in the archive the time variables in its decision vector is converted from seconds to days. This is necessary to make the archive human readable and also because the first thing the inner-loop does with an initial guess is convert from days to seconds, so it had better be in days when you pass it in!

git-svn-id: svn://svn.code.sf.net/p/emtg/emtg-goddard@33 77bb456d-b9d8-4424-8c4a-7c450a671a84