There are two bug fixes for the lazy race tree mode.  We weren't resetting the fail count on each new branch.
further, I've stitched in most of the MPI method for the lazy race tree.  It is entirely locked behind ifdef guards for the definition "EMTG_MPI".  I have confirmed that, without this guard, the code compiles and reduces to the original lazy race tree.

As for the MPI branch, I'm not sure the exit conditions have been set properly.

git-svn-id: svn://svn.code.sf.net/p/emtg/emtg-goddard@114 77bb456d-b9d8-4424-8c4a-7c450a671a84