Fix for the asteroid filter.  Apparently even though the asteroid bodies start at 1, the indicies still start at 0 in the list (go figure).  So although we've shifted all the asteroid identifiers up by 1 for the Universe file... we haven't actually shifted them up in the Universe.bodies list internal to the code.
I've adjusted this in the Asteroid filter, but I am not 100% sure if we need to make a similar adjustment to our destination list entry.  It *looks* like we DO NOT, based on searching through the code for all instances of options.sequence[x]  it seems to always be implemented as options.sequence[x-1] when it actually looks up a body inside mission/phase/problem, but I could be missing the fundamental case.

git-svn-id: svn://svn.code.sf.net/p/emtg/emtg-goddard@123 77bb456d-b9d8-4424-8c4a-7c450a671a84