Alex's new snoptProblemExtension and expected conversions of EMTG.  Does compile.  Has not been run-tested.
In all ways, the snoptProblemExtension should look, feel, and act just like snoptProblem that we had modified.  The only difference is that previously EMTG assumed that neA and neG were public; they have been returned to the correct private domain and a new getNeA() and getNeG() method is provided.

Note: snoptProblemExtension requires the original snopt.hh and snoptproblem.hh headers, as well as access to a compiled version of both snopt and the cppsrc directory.

Your code does not need to call either snopt.hh or snoptproblem.hh directly as snoptproblemextension should pass this through automatically.  Your code *does* need the /cppsrc directory and the /lib directory of snopt on your include and library path respectively.

git-svn-id: svn://svn.code.sf.net/p/emtg/emtg-goddard@246 77bb456d-b9d8-4424-8c4a-7c450a671a84