Merge from -unitscrubbing branch.
Branch began as an attempt to fix a few bugs as a feature branch, and quickly spiraled into a new trunk line accidently.

Primary additions from this branch:
- normalization of units everywhere
- NSGAII algorithm fixed/finalized
- new permanently added lambert solver
- new kepler solver
- numerous pyEMTG updates
- mildly improved error handling (aka a few more try/catch statements)
- new SNOPTProblemExtension interface
- a number of improved/more specific compiler-time switches
- a (new) makefile

note: this branch does not properly 'clean up' the compiler/installation issues.  That feature is coming soon, but for now, the included makefile needs to be hand-edited by every potential user, and the attached Visual Studio files are unlikely to work on any machine out-of-the-box unless you are Jacob.




git-svn-id: svn://svn.code.sf.net/p/emtg/emtg-goddard@249 77bb456d-b9d8-4424-8c4a-7c450a671a84