Successfully implemented derivatives with respect to previous time variables.
There is a strange issue with the derivative of the left-hand mass defect constraints with respect to previous step's state variables. Both SNOPT's finite differencing and mine say that derivative should be zero. However the mass defect constraint should be dependent on the mass flow rate in the previous time-step which in turn should be dependent on distance from the sun, which in turn is dependent on the state at the left-hand side of the previous state. I need to think some more.

git-svn-id: svn://svn.code.sf.net/p/emtg/emtg-goddard@342 77bb456d-b9d8-4424-8c4a-7c450a671a84