Upload of Matt's latest improvement to NSGA-II, with his comments:
"It turns out that elitism preservation in nominal NSGA-II is not leak-proof for more than two objectives.  A population can lose a boundary individual from generation to generation if the entire population is along the non-dominated front and there are multiple individuals with the same minimum objective function value (i.e., there is a boundary group, not single boundary individuals).  The algorithm only ensures that a single boundary individual will be retained via crowding distance assignment.  In our EMTG current code the individual that will be given the preferential crowding distance is just the first in the index from a sort along a particular objective (indifferent to the value of the other objectives).  The other individuals in the group have the nominal crowding distance assignment.  The problem is that in reality some boundary individuals are better than others, that is there are individuals that can dominate other individuals of the boundary group.  This is not accounted for in the Deb’s papers that I’ve referenced (early 2000s).
 
I modified the code so that when there are boundary groups for a given objective, the individuals in those groups with the best corresponding fitness for the other objectives will be given preference in the tournament operator.  This approach should ensure that the best individuals won’t be lost.  Another option is to do a non-dominated sort and assign crowding distance based on that rank across the entire group.  I’m not sure, which is most appropriate.  A case could be made that all the individuals in a boundary group should be given preferential crowding distance values, the trick is the population might not move if there are only a few possible objective values.
 
The likelihood of this issue cropping up is a function of the number objectives and type of objective.  If you are optimizing TOF, final mass, and C3 it is less likely to be an issue than when you are minimizing the number of thrusters and there are only 4 options (as an example).  Same story for more objectives – the more objectives the faster the entire population evolves to a single front."

git-svn-id: svn://svn.code.sf.net/p/emtg/emtg-goddard@358 77bb456d-b9d8-4424-8c4a-7c450a671a84