Made modifications to Astrodynamics.h and engine_model.cpp so that thruster coefficients are accessible from outside the find_engine_parameters() function so that GMATScripter can write out parameters for the built-in thruster models.
missionoptions.h now has a field for setting GMAT shadow model but there is not yet a way to turn it on and off from the GUI and/or script.

GMATScripter now output a script using the new GMAT propulsion and power models that is ingested properly by GMAT but I can't run it because the GMAT force model blows up. Is that the fault of the script or of the new GMAT features? Too early to say but this is good progress. (note: either way it's my fault...)

git-svn-id: svn://svn.code.sf.net/p/emtg/emtg-goddard@486 77bb456d-b9d8-4424-8c4a-7c450a671a84